backend/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.9.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
blinker==1.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
dj-database-url==3.0.1
Django==4.2.23
django-appconf==1.1.0
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
django-imagekit==5.0.0
django-js-asset==2.2.0
djangorestframework==3.16.0
docutils==0.20.1
Flask==3.0.0
frozenlist==1.3.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
gunicorn==23.0.0
idna==3.4
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.2.0
Kivy-Garden==0.1.5
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
magic-filter==1.0.9
MarkupSafe==2.1.3
multidict==6.0.4
packaging==25.0
pilkit==3.0
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.15.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyTelegramBotAPI==4.12.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2023.3
requests==2.30.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.5.3
tk==0.1.0
typing_extensions==4.14.1
urllib3==2.0.2
Werkzeug==3.0.1
whitenoise==6.9.0
yarl==1.9.2
zipp==3.17.0
